made changes to a couple of different files.
User login and registration works but only called upon clicking checkout
from within cart if the user has not been authenticated already. User
registration doesn't allow for duplicate emails. Upon checkout, the data
in the cart gets persisted to the Order and OrderDetail tables, and the
billing address into the Address table.

+++ src/main/java/com/estoremvc/controller/AdminController.java
@@ -163,7 +163,7 @@ public class AdminController {
 			break;
 		case user:
 			model = new ModelAndView("adminSearchUserPage");
-			item = userService.getUserById(id);
+			item = userService.getUserById((int)id);
 			model.addObject("user", (IUser)item);
 			break;
 		default:
